requirements.txt
Flask==2.2.2
flask-cors==3.0.10
python-dotenv==0.21.0
openai==0.27.0
